[BACKEND] Changes to contexts.py and views.py
Mainly changes to the spelling of "user_input" (formerly user-input) and changes to the context classes, so that the type is sent too.

Added contexts.py which includes representations of the context and user_input ASP syntax as well as methods to construct a list of options for each node which can be sent to the frontend. views.py changed to represent this.

NEEDS TO BE TESTED